New Model(Decision Tree) added! 1

+++ app.py
@@ -16,7 +16,7 @@ def predict():
     transaction_type = request.form.get('type')
 
     # Create the input array
-    input_query = np.array([[float(amount), int(time), int(date)]])
+    input_query = np.array([[int(amount), int(time), int(date)]])
 
     # Make the prediction
     result = model.predict(input_query)
